sd-632 introduce getUserContributedCampaigns
